Refactor tlp_type to conform to standards
Follows both "enum members are allcaps", and uses the names from the PCI-Express
manual. It also has refactoring of the logic to form TLPs from database rows so
it can be used in both "send" and "receive" modes. I'm aware that this is bad
commit hygeine. Guess how much I care.